✅ Fix RCMD check

+++ src/rcpp_find.h
@@ -1,6 +1,6 @@
 //#ifndef RCPP_FIND
 //#define RCPP_FIND
 
-int rcpp_find(int x, Rcpp::NumericVector y);
+int rcpp_find(double x, Rcpp::NumericVector y);
 
 //#endif // RCPP_FIND
